refactor: make multiplexers universal with templates
